use instance based partial schema
